Inhibit jump hooks when saving `point' on the first jump backward
On the first jump backward, `point' is saved to the jumplist so the jumplist
needs to be rotated once backward before the real backward jump the user
requested. Prevent the hooks from running when saving `point' is useful if the
hooks update some visual elements of the window, for example highlighting the
line after a jump. In the case of the first backward jump, the hook would be
run twice if they are not inhibited.