Fix incorrect setting of delete registers due to `kill-ring` length
The behavior of `(current-kill N)` is to always rotate to an element
within the `kill-ring` even if N is larger than the length of the
`kill-ring`.  If the length of the `kill-ring` is less than N, where N
is the Nth delete register, then setting the `car` of
`kill-ring-yank-pointer` will set the wrong register.