evil-jumps.el: Advise `delete-other-windows-internal`
Remove the advice on `delete-other-windows`.  That function may call
`delete-window` (which is already handled by the advice on
`delete-window-internal`) to do its work when there are some windows
that can't be deleted due to window parameters like `side` whereas the
work done in this advice assumes that all windows will be deleted.